+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandDescribe.cs
@@ -51,7 +51,13 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
             }
             else
             {
-                SetImpossible("Did not recognise parameter as a valid command");
+                // Maybe they typed the alias or name of a command
+                _nonDatabaseObjectToDescribe = new CommandInvoker(BasicActivator).GetSupportedCommands()
+                .FirstOrDefault(t=>BasicCommandExecution.HasCommandNameOrAlias(t,picker[0].RawValue));
+                
+                    
+                if(_nonDatabaseObjectToDescribe == null)
+                    SetImpossible("Did not recognise parameter as a valid command");
             }
         }
 
@@ -153,7 +159,8 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
             if (o is Type t)
             {
                 title = t.Name;
-                var docs = BasicActivator.CommentStore.GetTypeDocumentationIfExists(t, true, true);
+                var help = BasicActivator.CommentStore ?? CreateCommentStore();
+                var docs = help.GetTypeDocumentationIfExists(t, true, true);
                 return docs?.Trim() ?? "";
             }
 
@@ -340,11 +347,17 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
 
         private void PopulateBasicCommandInfo(StringBuilder sb,Type commandType)
         {
-            var help = new CommentStore();
-            help.ReadComments(Environment.CurrentDirectory);
+            var help = BasicActivator.CommentStore ?? CreateCommentStore();
 
             // Basic info about command
             sb.AppendLine("Name: " + commandType.Name);
+
+
+            var aliases = commandType.GetCustomAttributes(false).OfType<AliasAttribute>().ToArray();
+            if(aliases.Any())
+            {
+                sb.AppendLine("Aliases:" + string.Join(',',aliases.Select(a=>a.Name).ToArray()));
+            }
                 
             var helpText = help.GetTypeDocumentationIfExists(commandType);
 
@@ -364,5 +377,6 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
             sb.Append(" ");
 
         }
+
     }
 }