+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandListSupportedCommands.cs
@@ -5,27 +5,97 @@
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Text;
+using Rdmp.Core.Curation.Data;
+using ReusableLibraryCode.Checks;
 
 namespace Rdmp.Core.CommandExecution.AtomicCommands
 {
+    /// <summary>
+    /// Lists all known commands, optionally restricted to those matching pattern
+    /// </summary>
+    [Alias("lc")]
+    [Alias("ListCommands")]
     class ExecuteCommandListSupportedCommands:BasicCommandExecution
     {
-        public ExecuteCommandListSupportedCommands(IBasicActivateItems basicActivator):base(basicActivator)
+        private readonly string _pattern;
+        private readonly bool _verbose;
+
+        public ExecuteCommandListSupportedCommands(IBasicActivateItems basicActivator, 
+            [DemandsInitialization("Optional. A term to look for in command names.  Supports wildcards e.g. new*cata.  If not supplied then all will be shown")]
+            string pattern=null,
+            [DemandsInitialization("Optional. Set to true to display information about the command.  If specified with pattern then pattern will also search the description")]
+            bool verbose = false):base(basicActivator)
         {
-            
+            this._pattern = pattern;
+            this._verbose = verbose;
         }
 
         public override void Execute()
         {
             var commandCaller = new CommandInvoker(BasicActivator);
 
-            string commands = string.Join(Environment.NewLine,
-                commandCaller.GetSupportedCommands()
-                    .Select(t => BasicCommandExecution.GetCommandName(t.Name))
+            
+            var commands = commandCaller.GetSupportedCommands().ToArray();
+            var names = commands.Select(c=>BasicCommandExecution.GetCommandName(c.Name)).ToArray();
+            string[] descriptions;
+            
+             
+            if(_verbose)
+            {
+                var help = BasicActivator.CommentStore ?? CreateCommentStore();
+                descriptions = commands.Select(c=>help.GetTypeDocumentationIfExists(c)).ToArray();
+            }
+            else
+            {
+                descriptions = new string[commands.Length];
+            }
+
+            var onlyShowIndexes = new HashSet<int>();
+
+            if(!string.IsNullOrWhiteSpace(_pattern))
+            {
+                var tokens = _pattern.Split('*',StringSplitOptions.RemoveEmptyEntries);
+
+                for(int i=0;i<commands.Length;i++)
+                {
+                    if(tokens.All(t=>names[i].Contains(t,StringComparison.InvariantCultureIgnoreCase)))
+                    {
+                        onlyShowIndexes.Add(i);
+                    }
+                    if(tokens.All(t=>descriptions[i]?.Contains(t,StringComparison.InvariantCultureIgnoreCase) ?? false))
+                    {
+                        onlyShowIndexes.Add(i);
+                    }
+                }
+
+                // Nothing matches pattern
+                if(!onlyShowIndexes.Any())
+                {
+                    BasicActivator.GlobalErrorCheckNotifier.OnCheckPerformed(new CheckEventArgs("No commands matched supplied pattern",CheckResult.Warning));
+                    return;
+                }
+            }
+
+            bool showAll = onlyShowIndexes.Count == 0;
+            var outputCommandDictionary = new Dictionary<string,string>();
+
+            for(int i=0;i<commands.Length;i++)
+            {
+                if(showAll || onlyShowIndexes.Contains(i))
+                {
+                    var desc = string.IsNullOrWhiteSpace(descriptions[i]) ? "" : $"{Environment.NewLine}{descriptions[i]}{Environment.NewLine}";
+                    outputCommandDictionary.Add(names[i],desc);
+                }
+            }
+
+            string output = string.Join(Environment.NewLine,
+                    outputCommandDictionary.Select(kvp => kvp.Key+kvp.Value)
                         .OrderBy(s=>s));
 
-            BasicActivator.Show(commands);
+            BasicActivator.Show(output);
 
             base.Execute();
         }