+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandSet.cs
@@ -58,7 +58,7 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
         {
             _setOn = setOn;
 
-            _property = _setOn.GetType().GetProperty(property);
+            _property = GetProperty(_setOn, property);
 
             if(_setOn is IMightBeReadOnly m)
             {
@@ -95,6 +95,16 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
                     NewValue = picker[0].GetValueForParameterOfType(_property.PropertyType);
             }
         }
+
+        private PropertyInfo GetProperty(IMapsDirectlyToDatabaseTable setOn, string property)
+        {
+            var props = setOn.GetType().GetProperties();
+            
+            return props.FirstOrDefault(p => string.Equals(p.Name, property)) ??
+                   props.FirstOrDefault(p => string.Equals(p.Name, property,StringComparison.InvariantCultureIgnoreCase));
+
+        }
+
         public ExecuteCommandSet(IBasicActivateItems activator,IMapsDirectlyToDatabaseTable setOn,PropertyInfo property):base(activator)
         {
             _setOn = setOn;