+++ Rdmp.Core/CommandExecution/BasicCommandExecution.cs
@@ -22,6 +22,7 @@ using Rdmp.Core.DataViewing;
 using Rdmp.Core.Repositories.Construction;
 using ReusableLibraryCode;
 using ReusableLibraryCode.Checks;
+using ReusableLibraryCode.Comments;
 using ReusableLibraryCode.DataAccess;
 using ReusableLibraryCode.Icons.IconProvision;
 
@@ -509,5 +510,28 @@ namespace Rdmp.Core.CommandExecution
         {
             return GetCommandName();
         }
+
+        protected CommentStore CreateCommentStore()
+        {
+            var help = new CommentStore();
+            help.ReadComments(Environment.CurrentDirectory);
+            return help;
+        }
+
+        /// <summary>
+        /// Returns true if the supplied command Type is known (directly or via alias)
+        /// as <paramref name="name"/>
+        /// </summary>
+        public static bool HasCommandNameOrAlias(Type commandType, string name)
+        {
+            return 
+                    commandType.Name.Equals(BasicCommandExecution.ExecuteCommandPrefix + name,StringComparison.InvariantCultureIgnoreCase) 
+                    || 
+                    commandType.Name.Equals(name,StringComparison.InvariantCultureIgnoreCase) 
+                    || 
+                    commandType.GetCustomAttributes<AliasAttribute>(false)
+                    .Any(a=>a.Name.Equals(name,StringComparison.InvariantCultureIgnoreCase));
+        }
+
     }
 }