+++ Rdmp.Core/CommandLine/Options/ReleaseOptions.cs
@@ -17,13 +17,13 @@ namespace Rdmp.Core.CommandLine.Options
     public class ReleaseOptions : ConcurrentRDMPCommandLineOptions
     {
         [Option('c',"Configurations",HelpText = "List of ExtractionConfiguration IDs to release, they must all belong to the same Project")]
-        public IEnumerable<int> Configurations{ get; set; }
+        public string Configurations{ get; set; }
 
         [Option('p', "Pipeline", HelpText = "The ID of the release Pipeline to use")]
-        public int Pipeline { get; set; }
+        public string Pipeline { get; set; }
 
         [Option('s',"SelectedDatasets", HelpText = "List of SelectedDatasets IDs to release, they must all belong to ExtractionConfigurations within the same Project")]
-        public IEnumerable<int> SelectedDataSets { get; set; }
+        public string SelectedDataSets { get; set; }
 
         [Option('g', "Globals", HelpText = "True to release extracted globals (default) or false to skip them", Default = true)]
         public bool? ReleaseGlobals { get; set; }