+++ Tests.Common/Scenarios/TestsRequiringAnExtractionConfiguration.cs
@@ -148,8 +148,8 @@ namespace Tests.Common.Scenarios
 
             var runner = new ExtractionRunner(new ThrowImmediatelyActivator(RepositoryLocator),new ExtractionOptions()
             {
-                Command = CommandLineActivity.run, ExtractionConfiguration = _configuration.ID,
-                ExtractGlobals = true, Pipeline = pipeline.ID
+                Command = CommandLineActivity.run, ExtractionConfiguration = _configuration.ID.ToString(),
+                ExtractGlobals = true, Pipeline = pipeline.ID.ToString()
             });
 
             var returnCode = runner.Run(