+++ CHANGELOG.md
@@ -8,6 +8,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- Added checkbox for show/hide ProjectSpecific Catalogue columns in extraction configuration UI [#1265](https://github.com/HicServices/RDMP/issues/1265)
+- Integration tests and example scripts that can be run using RDMP command line `-f` option
+- The `Set` command no longer cares about property capitalization
+- Command line engines (e.g. `dle`) now optionally support specifying objects in command notation e.g. `-l "LoadMetadata:Load*Frank"` instead of `-l 1232`
+
+### Fixed
+
+- Fixed multiple calls to Refresh in DataExportPropertyManager causing Exception in extraction checks [#1274](https://github.com/HicServices/RDMP/issues/1274)
+- Fixed issues with Advanced column reorder dialog [#1273](https://github.com/HicServices/RDMP/issues/1273)
+  - Row size no longer cuts off bottom pixels of column name(s)
+  - Multi delete is now supported
+  - Pasted column name(s) with spaces e.g. `[my cool col]` now work
+- Fixed null reference in extraction checks when extracting a dataset where the original [ExtractionInformation] has been deleted [#1253](https://github.com/HicServices/RDMP/issues/1253)
+- Added an error provider message for when too many characters are entered in UIs with databindings [#1268](https://github.com/HicServices/RDMP/issues/1268).
+- Fixed running on command line with `-f somefile.yaml` being considered 'interactive' (i.e. RDMP would pause to ask you questions like 'are you sure?')
+- Fixed issue where DataTableUploadDestination pipeline component would refuse to load tables (e.g. from CSV) where the column has a full stop in it (e.g. `"mycol."`) [1269](https://github.com/HicServices/RDMP/issues/1269)
+
 ## [7.0.14] - 2022-06-27
 
 ### Added